requirements.txt
glfw==2.8.0
PyOpenGL==3.1.7
numpy==1.26.4
pillow==9.3.0